+++ darts/models/forecasting/__init__.py
@@ -54,6 +54,7 @@ PyTorch (Lightning)-based Models (`GlobalForecastingModel <https://unit8co.githu
     - :class:`~darts.models.forecasting.nf_model.NeuralForecastModel`
 Foundation Models (`GlobalForecastingModel <https://unit8co.github.io/darts/userguide/covariates.html#global-forecasting-models-gfms>`__)
     - :class:`~darts.models.forecasting.chronos2_model.Chronos2Model`
+    - :class:`~darts.models.forecasting.reverso_model.ReversoModel`
     - :class:`~darts.models.forecasting.timesfm2p5_model.TimesFM2p5Model`
 Ensemble Models (`GlobalForecastingModel <https://unit8co.github.io/darts/userguide/covariates.html#global-forecasting-models-gfms>`__)
     - :class:`~darts.models.forecasting.baselines.NaiveEnsembleModel`